move historylog to seperate file
